refactor(install): flatten install script layout
The install assets were nested under unix and windows directories even
though their filenames and extensions already identified the target
platform. Move the scripts and tests into contrib/install and rename
the test files so the layout stays shallow and the updated imports
remain correct after relocation.
